Fixed bug with duplicate entries in generated SRCINFO files
Any arrays that are used in the distro-specific release format with an 
architecture extension (i.e. 'focal_depends_x86_64') got duplicated 
inside of the generated SRCINFO file. This was due to a function looking 
for variables with the architecture extension, when it was only supposed 
to look for the base variable name ('focal_depends').