Fixed long generation times for SRCINFO files
Previously, a for loop that was used in SRCINFO generation was calling a 
command (which we check the stdout from) that took a long time to run 
multiple times. The output of that command only needs to be checked once 
though, so we now store it's value in a variable and pass it as an 
argument to all required functions.

Resolves https://github.com/makedeb/makedeb/issues/47