Fixed Docker images used for beta and stable deployments
Due to the dependencies used in the beta and stable releases, makedeb (or 'makedeb-alpha' rather) ends up getting uninstalled due to a dependency conflict. Using the appropriate image for the release we're building for alleviates the issue.

Also exited cleanly out of user repo publishing scripts when no files have been changed (happens when we don't update the package version, such as with this commit).