Added '-E' flag to sudo command used for build dependency installation in CI
The 'DEBIAN_FRONTEND' variable isn't being passed to apt due to sudo not passing the user's env to the commands it runs by default. Adding the '-E' argument fixes the issue.

Also added 'tzdata' directly to dependency installation command as it appears to be installing later in the script and thus causing a hang due to a prompt there (that isn't ignored by default due to the aformentioned 'DEBIAN_FRONTEND' issue).